refactor: migrate to typescript
