+++ .gitignore
@@ -3,5 +3,5 @@ node_modules
 package-lock.json
 .nyc_output
 
-test/types/*.js
-test/types/*.map
+*.js
+*.js.map
