+++ index.ts
@@ -1,0 +1,3 @@
+import SSHConfig from './src/ssh-config'
+export * from './src/ssh-config'
+export default SSHConfig
