+++ package.json
@@ -9,20 +9,23 @@
   },
   "files": [
     "index.js",
-    "src",
-    "types"
+    "src"
   ],
   "devDependencies": {
     "@types/mocha": "^9.1.0",
-    "@types/node": "^17.0.23",
-    "eslint": "^7.17.0",
+    "@types/node": "^17.0.45",
+    "@typescript-eslint/eslint-plugin": "^5.48.0",
+    "@typescript-eslint/parser": "^5.48.0",
+    "eslint": "^8.31.0",
     "heredoc": "^1.3.1",
     "mocha": "^8.2.1",
     "nyc": "^15.1.0",
     "typescript": "^4.6.3"
   },
   "scripts": {
-    "lint": "eslint .",
+    "lint": "eslint --ext ts .",
+    "lint:fix": "eslint --ext ts --fix .",
+    "prepack": "tsc",
     "pretest": "tsc",
     "test": "NODE_OPTIONS=--enable-source-maps mocha --exit --recursive",
     "test:coverage": "nyc mocha --exit --recursive && nyc report --reporter=lcov"
@@ -30,6 +33,5 @@
   "engine": {
     "node": ">= 10.0.0"
   },
-  "types": "types/index.d.ts",
   "license": "MIT"
 }
