+++ test/unit/ssh-config.test.ts
@@ -1,11 +1,7 @@
-'use strict'
-
-const assert = require('assert').strict
-const fs = require('fs')
-const path = require('path')
-const heredoc = require('heredoc').strip
-
-const SSHConfig = require('../..')
+import { strict as assert } from 'assert'
+import fs from 'fs'
+import path from 'path'
+import SSHConfig from '../..'
 
 const { DIRECTIVE } = SSHConfig
 
@@ -24,7 +20,7 @@ describe('SSHConfig', function() {
 
     // the first obtained parameter value will be used. So there's no way to
     // override parameter values.
-    assert(opts.ServerAliveInterval == 80)
+    assert.equal(opts.ServerAliveInterval, '80')
 
     // the computed result is flat on purpose.
     assert.deepEqual(config.compute('tahoe1'), {
@@ -44,11 +40,11 @@ describe('SSHConfig', function() {
   })
 
   it('.compute by Host with globbing', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host example*
         HostName example.com
         User simon
-    */}))
+    `)
 
     assert.deepEqual(config.compute('example1'), {
       Host: 'example*',
@@ -58,11 +54,11 @@ describe('SSHConfig', function() {
   })
 
   it('.compute by Host with multiple patterns', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host foo "*.bar" "baz ham"
         HostName example.com
         User robb
-    */}))
+    `)
 
     for (const host of ['foo', 'foo.bar', 'baz ham']) {
       assert.deepEqual(config.compute(host), {
@@ -79,10 +75,10 @@ describe('SSHConfig', function() {
    * - https://en.wikibooks.org/wiki/OpenSSH/Cookbook/Proxies_and_Jump_Hosts#Recursively_Chaining_an_Arbitrary_Number_of_Hosts
    */
   it('.compute by Host with chaining hosts', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host *+*
-        ProxyCommand ssh -W $(echo %h | sed 's/^.*+//;s/^\([^:]*$\)/\1:22/') $(echo %h | sed 's/+[^+]*$//;s/\([^+%%]*\)%%\([^+]*\)$/\2 -l \1/;s/:\([^:+]*\)$/ -p \1/')
-    */}))
+        ProxyCommand ssh -W $(echo %h | sed 's/^.*+//;s/^\\([^:]*$\\)/\\1:22/') $(echo %h | sed 's/+[^+]*$//;s/\\([^+%%]*\\)%%\\([^+]*\\)$/\\2 -l \\1/;s/:\\([^:+]*\\)$/ -p \\1/')
+    `)
 
     for (const host of ['host1+host2', 'host1:2022+host2:2224']) {
       const result = config.compute(host)
@@ -95,7 +91,7 @@ describe('SSHConfig', function() {
     const config = SSHConfig.parse(`
       Match host tahoe1
         HostName tahoe.com
-        
+
       Match host tahoe2
         HostName tahoe.org
     `)
@@ -108,7 +104,7 @@ describe('SSHConfig', function() {
     const config = SSHConfig.parse(`
       Match exec "return 1" host tahoe1
         HostName tahoe.com
-      
+
       Match exec "return 0" host tahoe1
         HostName tahoe.local
     `)
@@ -119,7 +115,6 @@ describe('SSHConfig', function() {
 
   it('.find with nothing shall yield error', async function() {
     const config = SSHConfig.parse(readFile('fixture/config'))
-    assert.throws(function() { config.find() })
     assert.throws(function() { config.find({}) })
   })
 
@@ -184,30 +179,28 @@ describe('SSHConfig', function() {
     assert(config.find({ Host: 'tahoe2' }) == null)
     assert(config.length === length - 1)
 
-    assert.throws(function() { config.remove() })
     assert.throws(function() { config.remove({}) })
   })
 
   it('.remove by function', async function() {
     const config = SSHConfig.parse(readFile('fixture/config'))
     const length = config.length
-    
-    config.remove((line) => line.param && line.param.toLowerCase() === 'Host'.toLowerCase() && line.value === 'tahoe2')
+
+    config.remove((line) => line.type === DIRECTIVE && /Host/i.test(line.param) && line.value === 'tahoe2')
     assert(config.find({ Host: 'tahoe2' }) == null)
     assert(config.length === length - 1)
 
-    assert.throws(function() { config.remove() })
     assert.throws(function() { config.remove({}) })
   })
 
   it('.append lines', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host example
         HostName example.com
         User root
         Port 22
         IdentityFile /path/to/key
-    */}))
+    `)
 
     config.append({
       Host: 'example2.com',
@@ -220,21 +213,21 @@ describe('SSHConfig', function() {
     assert.deepEqual(opts.IdentityFile, ['~/.ssh/id_rsa'])
     assert.deepEqual(config.find({ Host: 'example2.com' }), {
       type: DIRECTIVE,
-      before: '',
+      before: '      ',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'example2.com',
       config: new SSHConfig({
         type: DIRECTIVE,
-        before: '  ',
+        before: '        ',
         after: '\n',
         param: 'User',
         separator: ' ',
         value: 'pegg'
       },{
         type: DIRECTIVE,
-        before: '  ',
+        before: '        ',
         after: '\n',
         param: 'IdentityFile',
         separator: ' ',
@@ -244,13 +237,13 @@ describe('SSHConfig', function() {
   })
 
   it('.append with original identation recognized', function() {
-    const config = SSHConfig.parse(heredoc(function () {/*
+    const config = SSHConfig.parse(`
       Host example1
         HostName example1.com
         User simon
         Port 1000
         IdentityFile /path/to/key
-    */}).replace(/  /g, '\t'))
+    `.replace(/  /g, '\t'))
 
     config.append({
       Host: 'example3.com',
@@ -259,14 +252,14 @@ describe('SSHConfig', function() {
 
     assert.deepEqual(config.find({ Host: 'example3.com' }), {
       type: DIRECTIVE,
-      before: '',
+      before: '\t\t\t',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'example3.com',
       config: new SSHConfig({
         type: DIRECTIVE,
-        before: '\t',
+        before: '\t\t\t\t',
         after: '\n',
         param: 'User',
         separator: ' ',
@@ -276,22 +269,22 @@ describe('SSHConfig', function() {
   })
 
   it('.append with newline insersion', function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host test
-        HostName google.com*/}))
+        HostName google.com`)
 
     config.append({
       Host: 'test2',
       HostName: 'microsoft.com'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
+    assert.equal(config.toString(), `
       Host test
         HostName google.com
 
       Host test2
         HostName microsoft.com
-    */}))
+`)
   })
 
   it('.append to empty config', function() {
@@ -302,12 +295,11 @@ describe('SSHConfig', function() {
       HostName: 'example.com'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      IdentityFile ~/.ssh/id_rsa
+    assert.equal(config.toString(), `IdentityFile ~/.ssh/id_rsa
 
-      Host test2
-        HostName example.com
-    */}))
+Host test2
+  HostName example.com
+`)
   })
 
   it('.append to empty config with new section', function() {
@@ -317,10 +309,9 @@ describe('SSHConfig', function() {
       HostName: 'example.com',
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      Host test
-        HostName example.com
-    */}))
+    assert.equal(config.toString(), `Host test
+  HostName example.com
+`)
   })
 
   it('.append to empty section config', function() {
@@ -329,14 +320,13 @@ describe('SSHConfig', function() {
       HostName: 'example.com'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      Host test
-        HostName example.com
-    */}))
+    assert.equal(config.toString(), `Host test
+  HostName example.com
+`)
   })
 
   it('.compute with properties with multiple values', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host myHost
         HostName example.com
         LocalForward 1234 localhost:1234
@@ -347,7 +337,7 @@ describe('SSHConfig', function() {
 
       Host *
         CertificateFile /foo/bar3
-    */}))
+    `)
 
     assert.deepEqual(config.compute('myHost'), {
       Host: 'myHost',
@@ -359,13 +349,13 @@ describe('SSHConfig', function() {
   })
 
   it('.prepend lines', async function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host example
         HostName example.com
         User root
         Port 22
         IdentityFile /path/to/key
-    */}))
+    `)
 
     config.prepend({
       Host: 'examplePrepend2.com',
@@ -378,21 +368,21 @@ describe('SSHConfig', function() {
     assert.deepEqual(opts.IdentityFile, ['~/.ssh/id_rsa'])
     assert.deepEqual(config.find({ Host: 'examplePrepend2.com' }), {
       type: DIRECTIVE,
-      before: '',
+      before: '      ',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'examplePrepend2.com',
       config: new SSHConfig({
         type: DIRECTIVE,
-        before: '  ',
+        before: '        ',
         after: '\n',
         param: 'User',
         separator: ' ',
         value: 'pegg2'
       },{
         type: DIRECTIVE,
-        before: '  ',
+        before: '        ',
         after: '\n\n',
         param: 'IdentityFile',
         separator: ' ',
@@ -402,13 +392,13 @@ describe('SSHConfig', function() {
   })
 
   it('.prepend with original identation recognized', function() {
-    const config = SSHConfig.parse(heredoc(function () {/*
+    const config = SSHConfig.parse(`
       Host example1
         HostName example1.com
         User simon
         Port 1000
         IdentityFile /path/to/key
-    */}).replace(/  /g, '\t'))
+    `.replace(/  /g, '\t'))
 
     config.prepend({
       Host: 'examplePrepend3.com',
@@ -417,14 +407,14 @@ describe('SSHConfig', function() {
 
     assert.deepEqual(config.find({ Host: 'examplePrepend3.com' }), {
       type: DIRECTIVE,
-      before: '',
+      before: '\t\t\t',
       after: '\n',
       param: 'Host',
       separator: ' ',
       value: 'examplePrepend3.com',
       config: new SSHConfig({
         type: DIRECTIVE,
-        before: '\t',
+        before: '\t\t\t\t',
         after: '\n\n',
         param: 'User',
         separator: ' ',
@@ -434,21 +424,21 @@ describe('SSHConfig', function() {
   })
 
   it('.prepend with newline insertion', function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Host test
-        HostName google.com*/}))
+        HostName google.com`)
 
     config.prepend({
       Host: 'testPrepend2',
       HostName: 'microsoft.com'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      Host testPrepend2
+    assert.equal(config.toString(), `      Host testPrepend2
         HostName microsoft.com
 
+
       Host test
-        HostName google.com*/}))
+        HostName google.com`)
   })
 
   it('.prepend to empty config', function() {
@@ -459,13 +449,12 @@ describe('SSHConfig', function() {
       HostName: 'example.com'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      IdentityFile ~/.ssh/id_rsa
+    assert.equal(config.toString(), `IdentityFile ~/.ssh/id_rsa
 
-      Host prependTest
-        HostName example.com
+Host prependTest
+  HostName example.com
 
-      */}))
+`)
   })
 
   it('.prepend to empty config with new section', function() {
@@ -475,11 +464,10 @@ describe('SSHConfig', function() {
       HostName: 'example.com',
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      Host prependTest
-        HostName example.com
+    assert.equal(config.toString(), `Host prependTest
+  HostName example.com
 
-      */}))
+`)
   })
 
   it('.prepend to empty section config', function() {
@@ -489,70 +477,68 @@ describe('SSHConfig', function() {
       User: 'brian'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      HostName example.com
-      User brian
+    assert.equal(config.toString(), `HostName example.com
+User brian
 
-      Host test*/}))
+Host test`)
   })
 
   it('.prepend to empty section and existing section config', function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
-      Host test
+    const config = SSHConfig.parse(`
+Host test
 
-      Host test2
-        HostName google.com*/}))
+Host test2
+  HostName google.com`)
+
     config.prepend({
       HostName: 'example.com',
       User: 'brian'
     })
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      HostName example.com
-      User brian
+    assert.equal(config.toString(), `HostName example.com
+User brian
 
-      Host test
 
-      Host test2
-        HostName google.com*/}))
+Host test
+
+Host test2
+  HostName google.com`)
   })
 
   it('.prepend to with Include', function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
+    const config = SSHConfig.parse(`
       Include ~/.ssh/configs/*
 
       Host test2
-        HostName google.com*/}))
+        HostName google.com`)
 
     config.prepend({
       Host: 'example',
       HostName: 'microsoft.com',
     }, true)
 
-    assert.equal(config.toString(), heredoc(function() {/*
+    assert.equal(config.toString(), `
       Include ~/.ssh/configs/*
 
       Host example
         HostName microsoft.com
 
       Host test2
-        HostName google.com*/}))
+        HostName google.com`)
   })
 
   it('.prepend to with empty Include', function() {
-    const config = SSHConfig.parse(heredoc(function() {/*
-      Include ~/.ssh/configs/* */}))
+    const config = SSHConfig.parse('Include ~/.ssh/configs/* ')
 
     config.prepend({
       Host: 'example',
       HostName: 'microsoft.com',
     }, true)
 
-    assert.equal(config.toString(), heredoc(function() {/*
-      Include ~/.ssh/configs/*
+    assert.equal(config.toString(), `Include ~/.ssh/configs/*
 
-      Host example
-        HostName microsoft.com
-*/}))
+Host example
+  HostName microsoft.com
+`)
   })
 })
