+++ test/unit/stringify.test.ts
@@ -1,10 +1,9 @@
-'use strict'
 
-const assert = require('assert').strict || require('assert')
-const fs = require('fs')
-const heredoc = require('heredoc').strip
-const path = require('path')
-const SSHConfig = require('../..')
+import { strict as assert } from 'assert'
+import fs from 'fs'
+import path from 'path'
+import SSHConfig from '../..'
+
 const { parse, stringify } = SSHConfig
 
 function readFile(fname) {
@@ -20,53 +19,53 @@ describe('stringify', function() {
   })
 
   it('.stringify config with white spaces and comments retained', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       # Lake tahoe
       Host tahoe4
 
         HostName tahoe4.com
         # Breeze from the hills
         User keanu
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       # Lake tahoe
       Host tahoe4
 
         HostName tahoe4.com
         # Breeze from the hills
         User keanu
-    */}))
+    `)
   })
 
   it('.stringify IdentityFile entries with double quotes', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host example
         HostName example.com
         User dan
         IdentityFile "/path to my/.ssh/id_rsa"
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host example
         HostName example.com
         User dan
         IdentityFile "/path to my/.ssh/id_rsa"
-    */}))
+    `)
   })
 
   it('.stringify IndentityAgent entries with double quotes', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host example
         HostName example.com
         IdentityAgent "~/Library/Group Containers"
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host example
         HostName example.com
         IdentityAgent "~/Library/Group Containers"
-    */}))
+    `)
   })
 
   it('.stringify IndentityAgent entries with double quotes', function() {
@@ -76,66 +75,65 @@ describe('stringify', function() {
       IdentityAgent: '~/Library/Group Containers',
     })
 
-    assert.equal(stringify(config), heredoc(function() {/*
-      Host example
-        IdentityAgent "~/Library/Group Containers"
-    */}))
+    assert.equal(stringify(config), `Host example
+  IdentityAgent "~/Library/Group Containers"
+`)
   })
 
   it('.stringify Host entries with multiple patterns', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host foo bar  "baz"   "egg ham"
         HostName example.com
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host foo bar baz "egg ham"
         HostName example.com
-    */}))
+    `)
   })
 
   // #36
   it('.stringify User names with spaces', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host example
         User "dan abramov"
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host example
         User "dan abramov"
-    */}))
+    `)
   })
 
   // #38
   it('.stringify LocalForward without quotes', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host example
         LocalForward 1234 localhost:1234
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host example
         LocalForward 1234 localhost:1234
-    */}))
+    `)
   })
 
   it('.stringify multiple LocalForward', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host foo
         LocalForward 3128 127.0.0.1:3128
         LocalForward 3000 127.0.0.1:3000
-    */}))
+`)
 
     config.append({
       Host: 'bar',
       LocalForward: [
         '3128 127.0.0.1:3128',
         '3000 127.0.0.1:3000'
-      ]
+      ],
     })
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host foo
         LocalForward 3128 127.0.0.1:3128
         LocalForward 3000 127.0.0.1:3000
@@ -143,7 +141,7 @@ describe('stringify', function() {
       Host bar
         LocalForward 3128 127.0.0.1:3128
         LocalForward 3000 127.0.0.1:3000
-    */}))
+`)
   })
 
   // #43
@@ -153,23 +151,22 @@ describe('stringify', function() {
       IdentityFile: 'C:\\Users\\John Doe\\.ssh\\id_rsa'
     })
 
-    assert.equal(stringify(config), heredoc(function() {/*
-      Host foo
-        IdentityFile "C:\Users\John Doe\.ssh\id_rsa"
-    */}))
+    assert.equal(stringify(config), `Host foo
+  IdentityFile "C:\\Users\\John Doe\\.ssh\\id_rsa"
+`)
   })
 
   // https://github.com/microsoft/vscode-remote-release/issues/5562
   it('.stringify ProxyCommand with spaces', function() {
-    const config = parse(heredoc(function() {/*
+    const config = parse(`
       Host foo
         ProxyCommand "C:\foo bar\baz.exe" "arg" "arg" "arg"
-    */}))
+    `)
 
-    assert.equal(stringify(config), heredoc(function() {/*
+    assert.equal(stringify(config), `
       Host foo
         ProxyCommand "C:\foo bar\baz.exe" arg arg arg
-    */}))
+    `)
   })
 
   it('.stringify Match with criteria', function() {
